Feature/status http

+++ collecta/src/main/java/collectiva/org/collecta/service/ComentarioDoadorService.java
@@ -23,18 +23,27 @@ public class ComentarioDoadorService {
 
     public ResponseEntity<List<ComentarioDoador>> buscarTodosComentarios() {
         List<ComentarioDoador> comentarioDoador = comentarioDoadorRepository.findAll();
+        if (comentarioDoador.isEmpty()) {
+            return ResponseEntity.noContent().build();
+        }
         return ResponseEntity.ok().body(comentarioDoador);
     }
 
+
     public ResponseEntity<Optional<ComentarioDoador>> buscarComentarioPorId(UUID id) {
         Optional<ComentarioDoador> comentarioDoador = comentarioDoadorRepository.findById(id);
+        if (comentarioDoador.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         return ResponseEntity.ok().body(comentarioDoador);
     }
 
     public ResponseEntity<ComentarioDoador> atualizarComentario(UUID id, ComentarioDoador comentarioDoador) {
         Optional<ComentarioDoador> comentarioDoadorAntigo = comentarioDoadorRepository.findById(id);
+        if (comentarioDoadorAntigo.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         ComentarioDoador comentarioDoadorExistente = comentarioDoadorAntigo.get();
-
         ComentarioDoador comentarioDoadorAtualizado = ComentarioDoador.builder()
                 .id(comentarioDoadorExistente.getId())
                 .comentario(comentarioDoador.getComentario())
@@ -47,8 +56,10 @@ public class ComentarioDoadorService {
     }
 
     public ResponseEntity<Void> deletarComentario(UUID id) {
+        if (!comentarioDoadorRepository.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
         comentarioDoadorRepository.deleteById(id);
-        return ResponseEntity.ok().build();
+        return ResponseEntity.noContent().build();
     }
 }
-
