+++ collecta/src/main/java/collectiva/org/collecta/service/DoadorService.java
@@ -23,18 +23,26 @@ public class DoadorService {
 
     public ResponseEntity<List<Doador>> buscarTodosDoadores() {
         List<Doador> doador = doadorRepository.findAll();
+        if (doador.isEmpty()){
+            return ResponseEntity.noContent().build();
+        }
         return ResponseEntity.ok().body(doador);
     }
 
     public ResponseEntity<Optional<Doador>> buscarDoadorPorId(UUID id) {
         Optional<Doador> doador = doadorRepository.findById(id);
+        if (doador.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
         return ResponseEntity.ok().body(doador);
     }
 
     public ResponseEntity<Doador> atualizarDoador(UUID id, Doador doador) {
         Optional<Doador> doadorAntigo = doadorRepository.findById(id);
+        if (doadorAntigo.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
         Doador doadorExistente = doadorAntigo.get();
-
         Doador doadorAtualizado = Doador.builder()
                 .id(doadorExistente.getId())
                 .email(doador.getEmail())
@@ -52,6 +60,9 @@ public class DoadorService {
     }
 
     public ResponseEntity<Void> deletarDoador(UUID id) {
+        if (!doadorRepository.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
         doadorRepository.deleteById(id);
         return ResponseEntity.ok().build();
     }
